refactor: 상품 등록 서비스 함수 클래스 변경 (#CHZZ-35) - RegisterService -> AuctionService
