+++ src/main/java/org/chzz/market/domain/auction/controller/AuctionController.java
@@ -6,7 +6,6 @@ import org.chzz.market.domain.auction.dto.RegisterRequest;
 import org.chzz.market.domain.auction.dto.RegisterResponse;
 import org.chzz.market.domain.auction.dto.StartResponse;
 import org.chzz.market.domain.auction.service.AuctionService;
-import org.chzz.market.domain.auction.service.RegisterService;
 import org.chzz.market.domain.product.entity.Product.Category;
 import org.chzz.market.domain.auction.entity.SortType;
 import org.slf4j.Logger;
@@ -25,7 +24,6 @@ public class AuctionController {
     private static final Logger logger = LoggerFactory.getLogger(AuctionController.class);
 
     private final AuctionService auctionService;
-    private final RegisterService registerService;
 
     @GetMapping
     public ResponseEntity<?> getAuctionList(@RequestParam Category category,
@@ -40,19 +38,23 @@ public class AuctionController {
         return ResponseEntity.ok(auctionService.getAuctionDetails(auctionId, 1L)); // TODO: 추후에 인증된 사용자 정보로 수정 필요
     }
 
-    // 상품 등록
+    /**
+     * 상품 등록
+     */
     @PostMapping(value = "/register",
             consumes = {MediaType.MULTIPART_FORM_DATA_VALUE},
             produces = {MediaType.APPLICATION_JSON_VALUE})
     public ResponseEntity<RegisterResponse> createAuction(@ModelAttribute @Valid RegisterRequest request) {
 
-        RegisterResponse response = registerService.register(request);
+        RegisterResponse response = auctionService.register(request);
         logger.info("상품이 성공적으로 등록되었습니다. 상품 ID: {}", response.productId());
 
         return ResponseEntity.status(HttpStatus.CREATED).body(response);
     }
 
-    // 경매 상품으로 전환
+    /**
+     * 경매 상품으로 전환
+     */
     @PostMapping("/{auctionId}/start")
     public ResponseEntity<StartResponse> startAuction(@PathVariable Long auctionId, Long userId) {
         StartResponse response = auctionService.startAuction(auctionId, userId);
