--- src/main/java/org/chzz/market/domain/auction/service/RegisterService.java
@@ -1,9 +1,0 @@
-package org.chzz.market.domain.auction.service;
-
-
-import org.chzz.market.domain.auction.dto.RegisterRequest;
-import org.chzz.market.domain.auction.dto.RegisterResponse;
-
-public interface RegisterService {
-    RegisterResponse register(RegisterRequest request);
-}
